refactor: UserRepository를 domain에서 application 레이어로 이동
Clean Architecture 원칙에 따라 Repository 인터페이스를 적절한 레이어로 재배치:
- src/main/kotlin/com/deepromeet/atcha/user/domain/UserRepository.kt 삭제
- src/main/kotlin/com/deepromeet/atcha/user/application/UserRepository.kt 추가
- 사용자 탈퇴 기능을 위한 saveWithdrawalReason 메서드 추가
