--- src/main/kotlin/com/deepromeet/atcha/user/domain/UserRepository.kt
@@ -1,13 +1,0 @@
-package com.deepromeet.atcha.user.domain
-
-interface UserRepository {
-    fun findById(userId: UserId): User?
-
-    fun findByProviderId(providerId: String): User?
-
-    fun existsByProviderId(providerId: String): Boolean
-
-    fun save(user: User): User
-
-    fun delete(userId: UserId)
-}
